+++ Changelog.md
@@ -1,3 +1,5 @@
+# Development Version
+
 # 1.0.0
 
 * Added AMT practice to the UI, so that the experimenter can run the practice more easily ([#23](https://github.com/psychelzh/CAMP-imaging/issues/23)).