+++ Changelog.md
@@ -1,5 +1,7 @@
 # Development Version
 
+* Enhanced AMT task by setting stimuli size according to screen size.
+
 # 1.0.0
 
 * Added AMT practice to the UI, so that the experimenter can run the practice more easily ([#23](https://github.com/psychelzh/CAMP-imaging/issues/23)).
@@ -12,7 +14,6 @@
   * Added a wait-to-end screen after the experiment ([#19](https://github.com/CAMP-BNU/CAMP-imaging/issues/19)).
   * Added the real onset time of each trial to the response data.
   * Adjust the flip time to avoid stimulation early present
-  * Set stimuli size according to screen size
 
 # 0.2.1