+++ Changelog.md
@@ -1,6 +1,7 @@
 # Development Version
 
 * Enhanced AMT task by setting stimuli size according to screen size.
+* Fixed aspect ratio error for movie presentation.
 
 # 1.0.0